Use std::optional in STaskState
Requires moving to Client Deathmatch due to Game SA not supporting C++17
